pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "mas-aviary"
version = "1.0.0"
description = "Multi-agent system for aircraft design optimization via NASA Aviary"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Jessica Ezemba"}
]

dependencies = [
    "torch>=2.10.0",
    "transformers>=4.57.0",
    "smolagents>=1.24.0",
    "scikit-learn>=1.8.0",
    "matplotlib>=3.10.0",
    "pyyaml>=6.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.54.0", "altair>=6.0.0"]
vllm = ["openai>=2.24.0"]
tracking = ["wandb>=0.25.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=2.0",
]

[project.scripts]
mas-aviary = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests requiring GPU or real MCP (deselect with '-m not slow')",
    "gpu: marks tests requiring CUDA GPU",
    "mcp: marks tests requiring running MCP server",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
